File Route sebagai Peta Hak Akses

+++ PBL/backend_week6/app/service/student_authz_rules.go
@@ -5,20 +5,16 @@ import (
 	"latihan-fiber/helper"
 )
 
-// CanAccessStudent memutuskan apakah seseorang boleh menyentuh data student tertentu.
-// Memenuhi Tugas Mandiri C.2 poin 2: Fungsi murni tanpa import fiber/repository.
 func CanAccessStudent(
 	current model.AuthUser,
-	ownerID int64,
+	ownerID int, 
 	perms *helper.PermissionSet,
 	anyPermission string,
 ) bool {
-	// 1. Jalur Kepemilikan (Ownership): Data miliknya sendiri
-	// Karena UserID tipenya int dan ownerID int64, kita lakukan casting ke int64 agar tidak error.
-	if int64(current.UserID) == ownerID {
+	// Karena tipe datanya sekarang sudah sama-sama int, kita bisa langsung membandingkannya.
+	if current.UserID == ownerID {
 		return true
 	}
 
-	// 2. Jalur Hak Akses: Role-nya memang berhak atas data siapa pun (permission :any)
 	return perms.Can(current.Role, anyPermission)
 }
